Click btn and open page with form and comments
If you click CommentBtn you'll see page with tweet, form for adding new
comment and comments(if they exist). Comments are saved in the new collection.
"Comments". It contains postID, each of them contains comments field.
That is array and it contains every comment of this tweet. There has been
added a function "renderComments" in the OneTweetPage. Functional of That
fucntion is almost like renderPost in the Tweets component. If you add
new tweet, commentsCount will update.